passwordManager/requirements.txt
wtforms
Werkzeug==2.0.3
uuid
requests==2.27.1
python-dotenv==0.19.2
pyOpenSSL==22.0.0
PyJWT
flask-wtf
Flask-SQLAlchemy
flask-cors==3.0.10
flask_login
flask_bcrypt
Flask
email_validator
datetime
